Created unit tests
