+++ NuGet/Codaxy.WkHtmlToPdf/Codaxy.WkHtmlToPdf.nuspec
@@ -6,6 +6,6 @@
     <version>0.5.6</version>
     <authors>Codaxy</authors>
     <description>Convert HTML to PDF using wkhtmltopdf and C#</description>
-	<projectUrl>https://github.com/codaxy/wkhtmltopdf</projectUrl>	
+    <projectUrl>https://github.com/codaxy/wkhtmltopdf</projectUrl>
    </metadata>
 </package>
